+++ hive_flutter_io/lib/src/hive_extensions.dart
@@ -5,7 +5,7 @@ extension HiveX on HiveInterface {
   /// Initializes Hive with the path from [getApplicationDocumentsDirectory].
   ///
   /// You can provide a [subDir] where the boxes should be stored.
-  Future<void> initFlutter(String? subDir) async {
+  Future<void> initFlutter([String? subDir]) async {
     WidgetsFlutterBinding.ensureInitialized();
 
     String? path;